requirements.txt
django==1.9.2
pillow==3.1.1
django-photologue==3.5.1
django-envelope==1.1
django-crispy-forms==1.6
django-braces==1.8.1
